abstract fetching of both contracts into custom hook
